Create ringselect.py
Script to split flt files: One file which contains the peaks that belong to a certain phase. Another file which contains the leftover peaks (= all those peaks which certainly don't belong to that phase).
One has to keep in mind that for this script structure factors are not applied for the determination of the peak position of the phase. So all possible peak positions are considered.
Parts of the script (such as plotting and printing commands) are not (yet) useful for us but were part of the original script from Jon Wright so I left them in case we need them at some point.